Crate accumulator

This module provides algorithms for accessing and updating a Merkle Accumulator structure persisted in a key-value storage. Note that this doesn't write to the storage directly, rather, it reads from it via HashReader trait and yields writes via in memory HashMap.

Merkle Accumulator

Given an ever growing (append only) series of "leaf" hashes, we construct an evolving Merkle Tree with which proofs of inclusion/exclusion of a leaf hash at a leaf index in a snapshot of the tree (represented by root hash) can be given.

Leaf Nodes

Leaf nodes carry hash values to be stored and proved. They are only appended to the tree but never deleted or updated.

Internal Nodes

A non-leaf node carries the hash value derived from both its left and right children.

Dummy Nodes

To make sure each Leaf node has a Merkle Proof towards the root, dummy nodes are added so that along the route from a leaf to the root, each node has a sibling. Dummy nodes have the hash value of DEFAULT_DIGEST

A dummy node can appear as either a Leaf node or a non-Leaf node, but there is at most one dummy leaf at any time.

Frozen Nodes & Non-frozen Nodes

As leaves being added to the tree, dummy nodes get replaced by non-dummy nodes, and when a node has all it's descendants being non-dummy, it's "Frozen" -- its hash value won't change again in the event of new leafs being added. All leaves appended (not counting the one possible dummy leaf) are by definition Frozen.

Other nodes, which have one or more dummy descendants are Non-Frozen. As new elements are appended to the accumulator the hash value of these nodes will change.

Leaf Count

Given a count of the number of leaves in a Merkle Accumulator it is possible to determine the shape of the accumulator -- which nodes are filled and which nodes are dummy nodes.

Example: Logic view of a Merkle Accumulator with 5 leafs:

           Non-fzn
          /       \
         /         \
        /           \
      Fzn2         Non-fzn
     /   \           /   \
    /     \         /     \
   Fzn1    Fzn3  Non-fzn  [Dummy]
  /  \    /  \    /    \
 L0  L1  L2  L3 L4   [Dummy]

Position and Physical Representation

As a Merkle Accumulator tree expands to the right and upwards, we number newly frozen nodes monotonically. (One way to do it is simply to use in-order index of nodes.) What we call the above stated numbers identifying nodes: simply "Position".

And with that we can map a Merkle Accumulator into a key-value storage: key is the position of a node, value is hash value it carries.

We store only Frozen nodes, while generate non-Frozen nodes on the fly when accessing the tree. This way, the physical representation of the tree is append-only, i.e. once written to physical storage, nodes won't be either modified or deleted.

What we persist for the logical tree in the above example:

         Fzn2(3)
        /      \
       /        \
   Fzn1(1)       Fzn3(5)
  /     \       /     \
 L0(0)  L1(2)  L2(4)  L3(6)  L4(7)

To think in key-value pairs:

|<-key->|<--value-->|
|   0   | hash_L0   |
|   1   | hash_Fzn1 |
|   2   | hash_L1   |
|  ...  |   ...     |

Structs

MerkleAccumulator
In this live Merkle Accumulator algorithms.

Traits

HashReader
Defines the interface between MerkleAccumulator and underlying storage.