Struct accumulator::Accumulator

pub struct Accumulator<Key, Value> where Key: PartialOrd + Ord + Clone, Value: Clone {
    // some fields omitted
}

Implementation of Accumulator.

Methods

impl<Key: PartialOrd + Ord + Clone, Value: Clone> Accumulator<Key, Value>

fn with_capacity(quorum: usize, capacity: usize) -> Accumulator<Key, Value>

Constructor for capacity based Accumulator.

quorum defines the count at and above which add() will return Some().

fn with_duration(quorum: usize, duration: Duration) -> Accumulator<Key, Value>

Constructor for time based Accumulator.

quorum defines the count at and above which add() will return Some().

fn contains_key(&mut self, key: &Key) -> bool

Returns whether key exists in the accumulator or not.

fn is_quorum_reached(&mut self, key: &Key) -> bool

Returns whether key exists and has accumulated quorum or more corresponding values.

fn add(&mut self, key: Key, value: Value) -> Option<Vec<Value>>

Adds a key-value pair.

Returns the corresponding values for key if quorum or more values have been accumulated, otherwise returns None.

fn get(&mut self, key: &Key) -> Option<Vec<Value>>

Retrieves a clone of the values accumulated under key, or None if key doesn't exist.

fn delete(&mut self, key: &Key)

Removes key and all corresponding accumulated values.

fn cache_size(&mut self) -> usize

Returns the size of the accumulator, i.e. the number of keys held.

fn set_quorum_size(&mut self, new_size: usize)

Sets a new value for quorum.

This has immediate effect, even for existing key-value entries.