Crate accumulator

An accumulator container based on an Lru cach (time and size controlled) This container accumulate keys until a number of entries is reached. After this quaorum has been reached the container will continue to accept values for such keys this allows users to test merge functions until they are happy they have a good value. Otherwise a hacker could pass a single bad value and break all quorums

Structs

Accumulator
Accumulator for various message types