Crate accumulator

Re-exports

pub use utils::*;

Modules

utils

Structs

Accumulator

Represents a Universal Bilinear Accumulator.

Coefficient

A coefficent for updating witnesses

Delta

A compressed delta after evaluating the polynomials w.r.t an element

Element

An element in the accumulator

Error

An accumulator error

MembershipProof

A ZKP membership proof

MembershipProofCommitting

The commit or blinding step for generating a ZKP The next step is to call get_bytes_for_challenge to create the fiat shamir heuristic

MembershipProofFinal

The computed values after running MembershipProof.finalize

MembershipWitness

A membership witness that can be used for membership proof generation as described in section 4 in https://eprint.iacr.org/2020/777

NonMembershipProof

A ZKP non-membership proof

NonMembershipProofCommitting

The commit or blinding step for generating a ZKP The next step is to call get_bytes_for_challenge to create the fiat shamir heuristic

NonMembershipProofFinal

The computed values after running NonMembershipProof.finalize

NonMembershipWitness

A non-membership witness that can be used for non-membership proof generation as described in section 4 in https://eprint.iacr.org/2020/777

ProofParams

Section 8 in https://eprint.iacr.org/2020/777 setup calls for four distinct generators in G1

PublicKey

Represents \overline{Q} = \overline{P}*\alpha (public key) on page 6 in https://eprint.iacr.org/2020/777.pdf

SecretKey

Represents \alpha (secret key) on page 6 in https://eprint.iacr.org/2020/777.pdf

Enums

ProofMessage

The type of proof message. Either hidden or shared blinding Shared blinding is used to link to other proofs via Schnorr.

Functions

evaluate_delta

Computes the compressed delta needed to update a witness

evaluate_deltas

Compress the deltas for the specified element and return the single delta

pair

pairing

schnorr